Refine SAFA User Experience and Workflow
- Removed an external link from the SAFA banner for consistency.
- Included the configured base URL when representing the SafaConfig object for better visibility.
- Improved the delete project tool by adding a confirmation prompt and handling the case where the user has no projects to prevent accidental deletions.
- Streamlined the push commit process by automatically refreshing the project data, which includes running a summarization job on the latest version, after pushing the commits.
- Added the option to run a summarization job on the current project version during the refresh project process.
- Ensured that the search tool always creates a fresh vector store by deleting and recreating the vector store directory if it already exists.
- Added a utility function to delete a directory and its contents, which is used by the search tool to create a fresh vector store.
- Added a check in the input_menu_paged function to handle cases where the list of items is empty, preventing potential errors or unexpected behavior.

+++ safa/tools/projects/delete.py
@@ -1,5 +1,6 @@
 from safa.api.safa_client import SafaClient
 from safa.safa_config import SafaConfig
+from safa.utils.menus.inputs import input_confirm
 from safa.utils.menus.page_menu import input_menu_paged
 
 
@@ -11,10 +12,15 @@ def delete_project(config: SafaConfig, client: SafaClient) -> None:
     :return: None
     """
     projects = client.get_projects()
+    if len(projects) == 0:
+        print("User has no projects.")
+        return
     project_lookup_map = {p["name"]: p for p in projects}
-
-    project_name = input_menu_paged(list(project_lookup_map.keys()))
+    project_keys = list(project_lookup_map.keys())
+    project_name = input_menu_paged(project_keys)
     selected_project = project_lookup_map[project_name]
+    if not input_confirm(f"Are you sure you want to delete project ({project_name})?", default_value="y"):
+        return
     project_id = selected_project['projectId']
     client.delete_project(selected_project['projectId'])
     if project_id == config.project_id:
